refactor: other fixes and renaming
