+++ apps/frontend/src/views/EditorPage/EditorPage.tsx
@@ -1,11 +1,11 @@
 import { useState } from 'react';
 import EditorStyle from './EditorPage.module.css';
+import { components } from '@react-site-editor/ui';
 import SideBar from '@components/SideBar/SideBar';
 import SideBarHeader from '@components/SideBarHeader/SideBarHeader';
 import ComponentWrapper from '@components/ComponentWrapper/ComponentWrapper';
-import { components } from '@ui';
-import Preview from '@components/Preview/Preview';
-import Draggable from '@components/Decorator/Draggable';
+import PreviewIframe from '@components/Preview/PreviewIframe/PreviewIframe';
+import Draggable from '@components/Decorators/Draggable';
 
 const EditorPage: React.FunctionComponent = () => {
     const [selectedComponent, setSelectedComponent] = useState<string | null>(
@@ -40,7 +40,7 @@ const EditorPage: React.FunctionComponent = () => {
                 </ul>
             </SideBar>
             <div className={EditorStyle.preview}>
-                <Preview />
+                <PreviewIframe />
             </div>
             <SideBar scale="2">
                 <SideBarHeader>
@@ -52,6 +52,6 @@ const EditorPage: React.FunctionComponent = () => {
             </SideBar>
         </div>
     );
-};
+}
 
 export default EditorPage;
