+++ apps/frontend/src/views/Preview/Preview.tsx
@@ -1,10 +1,10 @@
 import { useCallback, useState } from 'react';
-import Dropable from '@components/Decorator/Dropable';
+import Droppable from '@components/Decorators/Droppable';
 import PreviewStyle from './Preview.module.css';
 
-const PreviewIframe: React.FunctionComponent = () => {
+const Preview: React.FunctionComponent = () => {
     const [result, setResult] = useState<string>('');
-    const [isHovered, setIsHover] = useState<boolean>(false);
+    const [isHovered, setIsHovered] = useState<boolean>(false);
 
     const defaultClassName =
         'text-black border-blue-500 flex justify-center p-4 border-2';
@@ -15,24 +15,24 @@ const PreviewIframe: React.FunctionComponent = () => {
 
     const handleDrop = (event: React.DragEvent) => {
         setResult(event.dataTransfer.getData('component'));
-        setIsHover(false);
+        setIsHovered(false);
     };
 
     const handleDragEnter = () => {
-        setIsHover(true);
+        setIsHovered(true);
     };
     const handleDragLeave = () => {
-        setIsHover(false);
+        setIsHovered(false);
     };
     return (
         <div className={PreviewStyle.iframe}>
-            <Dropable
+            <Droppable
                 onDrop={handleDrop}
                 onDragEnter={handleDragEnter}
                 onDragLeave={handleDragLeave}
                 type="component">
                 <div className={className()}>Drop it here</div>
-            </Dropable>
+            </Droppable>
             <div
                 dangerouslySetInnerHTML={{
                     __html: result
@@ -42,4 +42,4 @@ const PreviewIframe: React.FunctionComponent = () => {
     );
 };
 
-export default PreviewIframe;
+export default Preview;
